pyproject.toml
[tool.poetry]
name = "chmnuscheduleapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
django = "^5.0.3"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
django-ninja = "^1.1.0"
bcrypt = "^4.1.2"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
